+++ .gitignore
@@ -15,3 +15,6 @@ code-signing.pem.pub
 
 # mprocs
 mprocs.log
+
+# Module Federation generated types
+@mf-types/
