+++ packages/booking/package.json
@@ -14,45 +14,45 @@
     "check-deps": "rnx-align-deps"
   },
   "dependencies": {
-    "@bottom-tabs/react-navigation": "0.9.0",
-    "@module-federation/enhanced": "0.13.1",
-    "@react-native-async-storage/async-storage": "2.0.0",
-    "@react-navigation/native": "7.0.14",
-    "@react-navigation/native-stack": "7.2.0",
-    "react": "19.0.0",
-    "react-native": "0.78.2",
-    "react-native-bottom-tabs": "0.9.0",
+    "@bottom-tabs/react-navigation": "1.1.0",
+    "@module-federation/enhanced": "2.3.1",
+    "@react-native-async-storage/async-storage": "3.0.2",
+    "@react-navigation/native": "7.2.2",
+    "@react-navigation/native-stack": "7.14.10",
+    "react": "19.2.3",
+    "react-native": "0.84.1",
+    "react-native-bottom-tabs": "1.1.0",
     "react-native-calendars": "1.1291.1",
-    "react-native-edge-to-edge": "1.1.3",
-    "react-native-paper": "5.12.5",
-    "react-native-safe-area-context": "5.3.0",
-    "react-native-screens": "4.10.0",
-    "react-native-vector-icons": "10.2.0"
+    "react-native-edge-to-edge": "1.8.1",
+    "react-native-paper": "5.15.0",
+    "react-native-safe-area-context": "5.7.0",
+    "react-native-screens": "4.24.0",
+    "react-native-vector-icons": "10.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
     "@babel/preset-env": "^7.25.3",
     "@babel/runtime": "^7.25.0",
-    "@callstack/repack": "5.2.0",
-    "@react-native-community/cli": "15.0.1",
-    "@react-native/babel-preset": "0.78.2",
-    "@react-native/eslint-config": "0.78.2",
-    "@react-native/typescript-config": "0.78.2",
+    "@callstack/repack": "5.2.5",
+    "@react-native-community/cli": "20.1.0",
+    "@react-native/babel-preset": "0.84.1",
+    "@react-native/eslint-config": "0.84.1",
+    "@react-native/typescript-config": "0.84.1",
     "@rnx-kit/align-deps": "^2.5.1",
-    "@rspack/core": "1.5.0",
-    "@swc/helpers": "^0.5.17",
+    "@rspack/core": "1.7.11",
+    "@swc/helpers": "^0.5.20",
     "@types/jest": "^29.5.14",
-    "@types/react": "^19.0.0",
-    "@types/react-native-vector-icons": "^6.4.12",
-    "@types/react-test-renderer": "^19.0.0",
+    "@types/react": "^19.2.0",
+    "@types/react-native-vector-icons": "^6.4.18",
+    "@types/react-test-renderer": "^19.1.0",
     "@typescript-eslint/eslint-plugin": "^8.12.2",
     "@typescript-eslint/parser": "^8.12.2",
     "eslint": "^8.57.0",
     "jest": "^29.6.3",
     "prettier": "^2.8.8",
-    "react-test-renderer": "^19.0.0",
+    "react-test-renderer": "^19.2.3",
     "super-app-showcase-sdk": "0.0.2",
-    "typescript": "^5.6.3"
+    "typescript": "^5.8.3"
   },
   "rnx-kit": {
     "kitType": "app",
