+++ packages/booking/rspack.config.ts
@@ -14,16 +14,14 @@ const __dirname = path.dirname(__filename);
  * Learn about Re.Pack configuration: https://re-pack.dev/docs/guides/configuration
  */
 
-export default Repack.defineRspackConfig(({mode}) => {
+export default Repack.defineRspackConfig(({mode, platform}) => {
   return {
     mode,
     context: __dirname,
-    entry: {},
-    resolve: {
-      ...Repack.getResolveOptions(),
-    },
+    entry: './index.js',
+    resolve: {...Repack.getResolveOptions({enablePackageExports: true})},
     output: {
-      uniqueName: 'sas-auth',
+      uniqueName: 'sas-booking',
     },
     module: {
       rules: [
@@ -42,13 +40,15 @@ export default Repack.defineRspackConfig(({mode}) => {
     plugins: [
       new Repack.RepackPlugin(),
       new Repack.plugins.ModuleFederationPluginV2({
-        name: 'auth',
-        filename: 'auth.container.js.bundle',
-        dts: false,
+        name: 'booking',
+        filename: 'booking.container.js.bundle',
+        dts: true,
         exposes: {
-          './AccountScreen': './src/screens/AccountScreen',
-          './SignInScreen': './src/screens/SignInScreen',
-          './AuthProvider': './src/providers/AuthProvider',
+          './App': './src/navigation/MainNavigator',
+          './UpcomingScreen': './src/screens/UpcomingScreen',
+        },
+        remotes: {
+          auth: `auth@http://localhost:9003/${platform}/mf-manifest.json`,
         },
         shared: getSharedDependencies({eager: false}),
       }),
