+++ packages/booking/src/navigation/TabsNavigator.tsx
@@ -23,7 +23,9 @@ const TabsNavigator = () => {
     <Tabs.Navigator
       translucent={false}
       tabBarActiveTintColor={MD3Colors.primary50}
-      barTintColor={MD3Colors.primary95}>
+      tabBarInactiveTintColor={MD3Colors.primary95}
+      // barTintColor={MD3Colors.primary95}
+    >
       <Tabs.Screen
         name="HomeNavigator"
         component={HomeNavigator}
