+++ packages/host/android/app/src/main/AndroidManifest.xml
@@ -8,6 +8,7 @@
     android:roundIcon="@mipmap/ic_launcher_round"
     android:allowBackup="false"
     android:theme="@style/BootTheme"
+    android:usesCleartextTraffic="${usesCleartextTraffic}"
     android:supportsRtl="true">
     <activity
       android:name=".MainActivity"
