+++ packages/host/ios/Podfile
@@ -1,4 +1,9 @@
-require_relative '../node_modules/super-app-showcase-sdk/ios/react_native_setup.rb'
+# require_relative '../node_modules/super-app-showcase-sdk/ios/react_native_setup.rb'
+require Pod::Executable.execute_command('node', ['-p',
+  'require.resolve(
+    "react-native/scripts/react_native_pods.rb",
+    {paths: [process.argv[1]]},
+  )', __dir__]).strip
 
 platform :ios, min_ios_version_supported
 prepare_react_native_project!
@@ -21,7 +26,6 @@ target 'host' do
   pod 'SDWebImageSVGCoder', :modular_headers => true
 
   post_install do |installer|
-    # https://github.com/facebook/react-native/blob/main/packages/react-native/scripts/react_native_pods.rb#L197-L202
     react_native_post_install(
       installer,
       config[:reactNativePath],
