+++ packages/host/rspack.config.ts
@@ -20,10 +20,10 @@ export default Repack.defineRspackConfig(({mode, platform}) => {
     context: __dirname,
     entry: './index.js',
     resolve: {
-      ...Repack.getResolveOptions(),
+      ...Repack.getResolveOptions({enablePackageExports: true}),
     },
     output: {
-      uniqueName: 'sas-shopping',
+      uniqueName: 'sas-host',
     },
     module: {
       rules: [
@@ -36,26 +36,22 @@ export default Repack.defineRspackConfig(({mode, platform}) => {
           },
           type: 'javascript/auto',
         },
-        ...Repack.getAssetTransformRules({inline: true}),
+        ...Repack.getAssetTransformRules(),
       ],
     },
     plugins: [
       new Repack.RepackPlugin(),
       new Repack.plugins.ModuleFederationPluginV2({
-        name: 'shopping',
-        filename: 'shopping.container.js.bundle',
-        dts: false,
-        exposes: {
-          './App': './src/navigation/MainNavigator',
-        },
+        name: 'host',
+        dts: true,
         remotes: {
+          booking: `booking@http://localhost:9000/${platform}/mf-manifest.json`,
+          shopping: `shopping@http://localhost:9001/${platform}/mf-manifest.json`,
+          dashboard: `dashboard@http://localhost:9002/${platform}/mf-manifest.json`,
           auth: `auth@http://localhost:9003/${platform}/mf-manifest.json`,
+          news: `news@http://localhost:9004/${platform}/mf-manifest.json`,
         },
-        shared: getSharedDependencies({eager: false}),
-      }),
-      new Repack.plugins.CodeSigningPlugin({
-        enabled: mode === 'production',
-        privateKeyPath: path.join('..', '..', 'code-signing.pem'),
+        shared: getSharedDependencies({eager: true}),
       }),
       // silence missing @react-native-masked-view optionally required by @react-navigation/elements
       new rspack.IgnorePlugin({
