+++ packages/host/tsconfig.json
@@ -14,12 +14,17 @@
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "strict": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "paths": {
+      "*": ["./@mf-types/*"]
+    }
   },
+  "include": ["**/*.ts", "**/*.tsx", "./@mf-types/*"],
   "exclude": [
-    "node_modules",
     "babel.config.js",
     "metro.config.js",
-    "jest.config.js"
+    "jest.config.js",
+    "**/node_modules",
+    "**/Pods"
   ]
 }
