+++ packages/sdk/lib/dependencies.json
@@ -1,55 +1,55 @@
 {
   "react": {
     "name": "react",
-    "version": "19.0.0"
+    "version": "19.2.3"
   },
   "react-native": {
     "name": "react-native",
-    "version": "0.78.2"
+    "version": "0.84.1"
   },
   "@bottom-tabs/react-navigation": {
     "name": "@bottom-tabs/react-navigation",
-    "version": "0.9.0"
+    "version": "1.1.0"
   },
   "@react-native-async-storage/async-storage": {
     "name": "@react-native-async-storage/async-storage",
-    "version": "2.0.0"
+    "version": "3.0.2"
   },
   "@react-navigation/native": {
     "name": "@react-navigation/native",
-    "version": "7.0.14"
+    "version": "7.2.2"
   },
   "@react-navigation/native-stack": {
     "name": "@react-navigation/native-stack",
-    "version": "7.2.0"
+    "version": "7.14.10"
   },
   "react-native-bottom-tabs": {
     "name": "react-native-bottom-tabs",
-    "version": "0.9.0"
+    "version": "1.1.0"
   },
   "react-native-paper": {
     "name": "react-native-paper",
-    "version": "5.12.5"
+    "version": "5.15.0"
   },
   "react-native-safe-area-context": {
     "name": "react-native-safe-area-context",
-    "version": "5.3.0"
+    "version": "5.7.0"
   },
   "react-native-screens": {
     "name": "react-native-screens",
-    "version": "4.10.0"
+    "version": "4.24.0"
   },
   "react-native-vector-icons": {
     "name": "react-native-vector-icons",
-    "version": "10.2.0"
+    "version": "10.3.0"
   },
   "@module-federation/enhanced": {
     "name": "@module-federation/enhanced",
-    "version": "0.13.1",
+    "version": "2.3.1",
     "shared": false
   },
   "react-native-edge-to-edge": {
     "name": "react-native-edge-to-edge",
-    "version": "1.1.3"
+    "version": "1.8.1"
   }
 }
