+++ packages/shopping/rspack.config.ts
@@ -19,11 +19,9 @@ export default Repack.defineRspackConfig(({mode, platform}) => {
     mode,
     context: __dirname,
     entry: './index.js',
-    resolve: {
-      ...Repack.getResolveOptions(),
-    },
+    resolve: {...Repack.getResolveOptions({enablePackageExports: true})},
     output: {
-      uniqueName: 'sas-host',
+      uniqueName: 'sas-shopping',
     },
     module: {
       rules: [
@@ -36,22 +34,26 @@ export default Repack.defineRspackConfig(({mode, platform}) => {
           },
           type: 'javascript/auto',
         },
-        ...Repack.getAssetTransformRules(),
+        ...Repack.getAssetTransformRules({inline: true}),
       ],
     },
     plugins: [
       new Repack.RepackPlugin(),
       new Repack.plugins.ModuleFederationPluginV2({
-        name: 'host',
-        dts: false,
-        remotes: {
-          booking: `booking@http://localhost:9000/${platform}/mf-manifest.json`,
-          shopping: `shopping@http://localhost:9001/${platform}/mf-manifest.json`,
-          dashboard: `dashboard@http://localhost:9002/${platform}/mf-manifest.json`,
-          auth: `auth@http://localhost:9003/${platform}/mf-manifest.json`,
-          news: `news@http://localhost:9004/${platform}/mf-manifest.json`,
+        name: 'shopping',
+        filename: 'shopping.container.js.bundle',
+        dts: true,
+        exposes: {
+          './App': './src/navigation/MainNavigator',
         },
-        shared: getSharedDependencies({eager: true}),
+        remotes: {
+          auth: `auth@http://localhost:9003/${platform}/mf-manifest.json`,
+        },
+        shared: getSharedDependencies({eager: false}),
+      }),
+      new Repack.plugins.CodeSigningPlugin({
+        enabled: mode === 'production',
+        privateKeyPath: path.join('..', '..', 'code-signing.pem'),
       }),
       // silence missing @react-native-masked-view optionally required by @react-navigation/elements
       new rspack.IgnorePlugin({
