+++ packages/shopping/tsconfig.json
@@ -1,6 +1,12 @@
 {
-  "extends": "@react-native/typescript-config/tsconfig.json",
+  "extends": "@react-native/typescript-config",
   "compilerOptions": {
-    "module": "es2020"
-  }
+    "types": ["jest"],
+    "module": "es2020",
+    "paths": {
+      "*": ["./@mf-types/*"]
+    }
+  },
+  "include": ["**/*.ts", "**/*.tsx", "./@mf-types/*"],
+  "exclude": ["**/node_modules", "**/Pods"]
 }
